fix: prevent C level fatal compiler error
Due to a convuluted mix of returning a mutable reference but by using
the interfaces type signature it looks like V's compiler is getting all
flustered and confused. This isn't a bug with the V compiler perse, but
the normal nice error/compilation handling isn't taking care of this
case, so when I get time I really should be raising an issue with the V
team.