fix: prevent subsequent offsets interfering with clip areas
This fix basically ensures that a clip area's location and size are only
adjusted by the existing offsets in the stack present up until the clip
areas creation, and does not apply offsets after it was created. Without
this fix, we observe a clip areas position and size will drift and never
correctly apply to subsequent draws which are themselves using the
offset stack in full.